pyproject.toml
[tool.poetry]
name = "tracker_tool"
version = "0.0.0"
description = "A short description of the project"
authors = [
    "Jonas Gerber <j.gerber2803@gmail.com>",
]
readme = "README.md"
license = "MIT"
classifiers = [
    "Development Status :: 1 - Planning",
]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
hydra-core = "^1.1.1"
python-dotenv = "^0.21.0"
autopep8 = "^2.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-xdist = "^2.5.0"
pytest-cov = "^4.0.0"
pre-commit = "^2.17.0"
pdoc = "^7.1.1"
black = "^22.6.0"
isort = "^5.10.1"
readme-coverage-badger = ">=0.1.2,<1.0.0"
flake8 = "^6.0.0"
pylint = "^2.15.9"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = [
    '--verbose',
    '--durations=10',
    '--color=yes',
    '-s',
    '-vv',
    '--doctest-modules',
    '--cov=src/tracker_tool',
    '-n 2',
]
xfail_strict = true
filterwarnings = ["ignore"]
log_cli_level = "info"
testpaths = ["tests", "src/tracker_tool"]

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
	\.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| _build
| buck-out
| build
)/
'''

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
